+++ collective/venue/eventaccessor.py
@@ -29,15 +29,17 @@ class VenueEventAccessor(EventAccessor):
     @property
     def location(self):
         context = self.context
-        location_ref = ILocation(context)
+        location_ref = ILocation(context, None)
+        if not location_ref:
+            return
         location_uid = location_ref.location_uid
         location_notes = location_ref.location_notes
-        location_url = location_ref.location_url
         location = uuidToObject(location_uid)
 
         meta_basic = IBasic(location, None)
         add = IAddress(location, None)
 
+        location_url = None
         ret = u''
         if meta_basic and add:
             # I'm a location reference.