+++ CHANGES.rst
@@ -4,6 +4,15 @@ Changelog
 3.1.2 (unreleased)
 ------------------
 
+- Prevent ``ILocation`` behavior not adaptable in ``VenueEventAccessor``.
+  [thet]
+
+- Add ``IICalendarEventComponent`` adapter to customize the ical export of ``location``, ``contact`` and ``geo`` properties.
+  [thet]
+
+- Only include the leaflet bundle where it is needed via ``add_bundle_on_request``.
+  [thet]
+
 - Adapt the ``contenttype-venue`` icon to recent changes in plonetheme.barceloneta 1.7.4.
   [thet]